remove: NetworkSessionManager 코드 분리
